+++ familyhelper-plugin-assets/src/main/java/com/dwarfeng/familyhelper/plugin/assets/bean/dto/VoucherIdWrapper.java
@@ -1,0 +1,44 @@
+package com.dwarfeng.familyhelper.plugin.assets.bean.dto;
+
+import com.dwarfeng.subgrade.stack.bean.dto.Dto;
+import org.jboss.resteasy.annotations.jaxrs.FormParam;
+import org.jboss.resteasy.annotations.providers.multipart.PartType;
+
+import javax.ws.rs.core.MediaType;
+
+/**
+ * 凭证 ID 包装器。
+ *
+ * @author DwArFeng
+ * @since 1.4.0
+ */
+public class VoucherIdWrapper implements Dto {
+
+    private static final long serialVersionUID = 1844921359119691612L;
+
+    @FormParam("voucher-id")
+    @PartType(MediaType.TEXT_PLAIN)
+    private Long voucherId;
+
+    public VoucherIdWrapper() {
+    }
+
+    public VoucherIdWrapper(Long voucherId) {
+        this.voucherId = voucherId;
+    }
+
+    public Long getVoucherId() {
+        return voucherId;
+    }
+
+    public void setVoucherId(Long voucherId) {
+        this.voucherId = voucherId;
+    }
+
+    @Override
+    public String toString() {
+        return "VoucherIdWrapper{" +
+                "voucherId=" + voucherId +
+                '}';
+    }
+}